+++ .vuepress/enhanceApp.js
@@ -35,7 +35,8 @@ const iconsToMock = [
   'faLightbulb',
   'faThumbsUp',
   'faEllipsisV',
-  'faTimes'
+  'faTimes',
+  'faCheck'
 ]
 const mockedFalIcons = _.mapValues(_.pick(fas, iconsToMock), function (original) {
   return _.assign({}, original, {